feat: added generic handler for webhook handler
