+++ server/entities/payment/controllers/handle-payment-webhook.js
@@ -1,0 +1,16 @@
+import { paymentLogger } from '../../../utils'
+
+export const handlePaymentWebhook = async (req, res) => {
+   try {
+      const paymentType = 'stripe'
+      const functionFilePath = `../functions/${paymentType}`
+      const method = await import(functionFilePath)
+      const result = await method.default(req, 'webhook')
+      if (result.success && result.data) {
+         await paymentLogger(result.data)
+      }
+      return res.status(result.code).json(result)
+   } catch (error) {
+      return res.status(500).json({ success: false, error })
+   }
+}
