+++ server/entities/payment/controllers/index.js
@@ -1,5 +1,5 @@
 export * from './initiate-payment'
-export * from './stripe-webhook'
+export * from './handle-payment-webhook'
 export * from './setup-intent'
 export * from './handle-cart-payment'
 export * from './payment-method'
